cloud/requirements.txt
fastparquet==2024.5.0
minio==7.1.16
PyYAML==6.0.1
pandas==3.0.1
pyarrow==23.0.1
pyspark==4.1.1
pydeequ==1.5.0
python-dotenv==1.2.2
psycopg2-binary==2.9.11
google-cloud-storage==3.9.0
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.1
airflow-provider-great-expectations==0.2.6
apache-airflow-providers-docker==3.7.5
ansible==12.3.0